% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add.hopar.R
\name{add.hopar}
\alias{add.hopar}
\title{Add prey information to a dataframe}
\usage{
add.hopar(faeduhopur, flokkur, col.names = c("thyngd"), isl = F,
  return.on.error = F, hopur = NULL, tmptable = F,
  oracle = fjolstOracle, hopar.col = hopar.col)
}
\arguments{
\item{faeduhopur}{xxx}

\item{flokkur}{xxx}

\item{col.names}{xxx}

\item{isl}{xxx}

\item{return.on.error}{xxx}

\item{hopur}{xxx}

\item{tmptable}{xxx}

\item{oracle}{xxx}

\item{hopar.col}{xxx}
}
\description{
Add prey information to a dataframe
}

