% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/afli.per.stod.R
\name{afli.per.stod}
\alias{afli.per.stod}
\title{catch per station}
\usage{
afli.per.stod(sfile, lwcoeff = c(0.01, 3), teg = 1, lfile, minl, maxl,
  oracle = fjolstOracle, name = "afli")
}
\arguments{
\item{sfile}{xxx}

\item{lwcoeff}{xxx}

\item{teg}{xxx}

\item{lfile}{xxx}

\item{minl}{xxx}

\item{maxl}{xxx}

\item{oracle}{xxx}

\item{name}{xxx}
}
\description{
catch per station
}

