% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{apply.shrink}
\alias{apply.shrink}
\title{Apply a function to a vector for a combination of categories.}
\usage{
apply.shrink(X, INDICES, FUN = NULL, names, ...)
}
\arguments{
\item{X}{Input data to \code{FUN}}

\item{INDICES}{list of categories to be combined}

\item{FUN}{Function to be applied}

\item{names}{Column names for the resulting dataframe}

\item{\dots}{Additional arguments to \code{FUN}}
}
\value{
Dataframe of outcomes applying \code{FUN} to \code{X} for the
combination of categories in \code{INDICES}
}
\description{
Apply a function to a vector for a combination of categories.
}
\note{
Needs elaboration, or could be dropped/hidden, use merge instead.
}
\seealso{
\code{\link{apply.shrink.dataframe}}, \code{\link{merge}}
}
\keyword{manip}

