% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faeduhopar}
\alias{faeduhopar}
\title{List of all food items in the faeda database}
\description{
Name, icelandic name, group (fish, etc), english name and species code for
each group.
}

