% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fjoldi.a.fstodvum.R
\name{fjoldi.a.fstodvum}
\alias{fjoldi.a.fstodvum}
\title{Calculate number in each lengthgroup at some stations.  (out of date
function)}
\usage{
fjoldi.a.fstodvum(fstodvar, teg)
}
\arguments{
\item{fstodvar}{xxx}

\item{teg}{xxx}
}
\description{
Calculate number in each lengthgroup at some stations.  (out of date
function)
}

