% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fjoldi.i.flokkum.R
\name{fjoldi.i.flokkum}
\alias{fjoldi.i.flokkum}
\title{Calculate number of fish behind (not in) each food sample.}
\usage{
fjoldi.i.flokkum(flokkur, teg, option = 1, lengdir, numer,
  lenfl = old.stomach.breaks$breaks, oracle = fjolstOracle)
}
\arguments{
\item{flokkur}{xxx}

\item{teg}{xxx}

\item{option}{xxx}

\item{lengdir}{xxx}

\item{numer}{xxx}

\item{lenfl}{xxx}

\item{oracle}{xxx}
}
\description{
Calculate number of fish behind (not in) each food sample.
}

