% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fjperstod.R
\name{fjperstod}
\alias{fjperstod}
\title{Calculate the number of a species at the list of stations.}
\usage{
fjperstod(data, tegund, name = "fjoldi", oracle = fjolstOracle)
}
\arguments{
\item{data}{xxx}

\item{tegund}{xxx}

\item{name}{xxx}

\item{oracle}{xxx}
}
\description{
Calculate the number of a species at the list of stations.
}

