% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flokkar.col}
\alias{flokkar.col}
\title{Selected columns from the table f_flokkar.}
\format{The format is: chr [1:12] "synis.id.id" "flokk.id" "ranfiskur"
"lenfl" ...}
\description{
Selected columns from the table f_flokkar.
}

