% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hopar.col}
\alias{hopar.col}
\title{Names of columns in faeda.f_hopar}
\format{The format is: chr [1:5] "flokk.id" "faeduhopur" "fjoldi" "thyngd"
"melt"}
\description{
Names of columns in faeda.f_hopar
}

