% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\title{The function join (also in the geo library).  SImilar to merge.}
\usage{
join(x, y, column, name.x, name.y, outer.join = T, set = NA)
}
\arguments{
\item{x}{xxx}

\item{y}{xxx}

\item{column}{xxx}

\item{name.x}{xxx}

\item{name.y}{xxx}

\item{outer.join}{xxx}

\item{set}{xxx}
}
\description{
The function join (also in the geo library).  SImilar to merge.
}

