% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kgpertog.R
\name{kgpertog}
\alias{kgpertog}
\title{calculate catch ber station}
\usage{
kgpertog(sfile, lwcoeff = c(0.01, 3), teg = 1, lfile, minl, maxl,
  oracle = fjolstOracle, name = "afli")
}
\arguments{
\item{sfile}{xxx}

\item{lwcoeff}{xxx}

\item{teg}{xxx}

\item{lfile}{xxx}

\item{minl}{xxx}

\item{maxl}{xxx}

\item{oracle}{xxx}

\item{name}{xxx}
}
\description{
calculate catch ber station
}

