% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kvarnir.col.oracle}
\alias{kvarnir.col.oracle}
\title{Name of columns in the oracle tables fiskar.kvarnir.}
\format{The format is: chr [1:23] "synis_id" "tegund" "nr" "lengd" "kyn"
"kynthroski" ...}
\description{
Name of columns in the oracle tables fiskar.kvarnir.
}

