% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kvarnir.col}
\alias{kvarnir.col}
\title{Splus or R correspondence of kvarnir.col.oracle.  "_ "replaced by "." .}
\format{The format is: chr [1:23] "synis.id" "tegund" "nr" "lengd" "kyn"
"kynthroski" ...}
\description{
Splus or R correspondence of kvarnir.col.oracle.  "_ "replaced by "." .
}

