% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lengdir.col.oracle}
\alias{lengdir.col.oracle}
\title{Name of columns in the oracle tables fiskar.lengdir}
\format{The format is: chr [1:6] "synis_id" "tegund" "lengd" "fjoldi" "kyn"
...}
\description{
Name of columns in the oracle tables fiskar.lengdir
}

