% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lengdir.col}
\alias{lengdir.col}
\title{Splus or R correspondence of lengdir.col.oracle.  "_ "replaced by "." .}
\format{The format is: chr [1:6] "synis.id" "tegund" "lengd" "fjoldi" "kyn"
...}
\description{
Splus or R correspondence of lengdir.col.oracle.  "_ "replaced by "." .
}

