% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lesa.allir.hopar.R
\name{Lesa.allir.hopar}
\alias{Lesa.allir.hopar}
\title{Read all records from the table f_hopar corresponding to the selected
samples.}
\usage{
Lesa.allir.hopar(flokk.id, tmptable = F, oracle = fjolstOracle)
}
\arguments{
\item{flokk.id}{xxx}

\item{tmptable}{xxx}

\item{oracle}{xxx}
}
\description{
Read all records from the table f_hopar corresponding to the selected
samples.
}

