% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lesa.f.lengdir.R
\name{lesa.f.lengdir}
\alias{lesa.f.lengdir}
\title{Read all length measurement of selected prey in selected predator at given
stations.}
\usage{
lesa.f.lengdir(synis.id = NULL, ranfiskur, lenfl = NULL, faeduhopur,
  table = "f_lengdir", haed = F, kyn = F, faerslunumer = T,
  oracle = fjolstOracle)
}
\arguments{
\item{synis.id}{xxx}

\item{ranfiskur}{xxx}

\item{lenfl}{xxx}

\item{faeduhopur}{xxx}

\item{table}{xxx}

\item{haed}{xxx}

\item{kyn}{xxx}

\item{faerslunumer}{xxx}

\item{oracle}{xxx}
}
\description{
Read all length measurement of selected prey in selected predator at given
stations.
}

