% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lesa.flokkar.R
\name{lesa.flokkar}
\alias{lesa.flokkar}
\title{Read from the table faeda.f_flokkar or the dataframe fflokkar.}
\usage{
lesa.flokkar(ranfiskur, lenfl = NULL, synis.id = NULL, fjoldi = T,
  total = T, flokk.id = NULL, oracle = fjolstOracle,
  old.stomach.breaks = old.stomach.breaks)
}
\arguments{
\item{ranfiskur}{xxx}

\item{lenfl}{xxx}

\item{synis.id}{xxx}

\item{fjoldi}{xxx}

\item{total}{xxx}

\item{flokk.id}{xxx}

\item{oracle}{xxx}

\item{old.stomach.breaks}{xxx}
}
\description{
Read from the table faeda.f_flokkar or the dataframe fflokkar.
}

