% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lesa.hopar.R
\name{lesa.hopar}
\alias{lesa.hopar}
\title{Read from the table faeda.f_hopar or the dataframe fhopar}
\usage{
lesa.hopar(faeduhopur, flokk.id = NULL, col.names = NULL,
  return.on.error = F, tmptable = F, oracle = fjolstOracle)
}
\arguments{
\item{faeduhopur}{xxx}

\item{flokk.id}{xxx}

\item{col.names}{xxx}

\item{return.on.error}{xxx}

\item{tmptable}{xxx}

\item{oracle}{xxx}
}
\description{
Read from the table faeda.f_hopar or the dataframe fhopar
}

