% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lesa.kvarnir.R
\name{lesa.kvarnir}
\alias{lesa.kvarnir}
\title{Read from the table fiskar.kvarnir or the dataframe all.kv.}
\usage{
lesa.kvarnir(synis.id, teg, col.names = c("synis.id", "aldur", "lengd"),
  aldur = NULL, kyn = NULL, kynthroski = NULL, oslaegt = F,
  slaegt = F, lifur = F, kynfaeri = F, lengd = NULL, leidrett = F,
  oracle = fjolstOracle, kvarnir.col = kvarnir.col,
  kvarnir.col.oracle = kvarnir.col.oracle)
}
\arguments{
\item{synis.id}{xxx}

\item{teg}{xxx}

\item{col.names}{xxx}

\item{aldur}{xxx}

\item{kyn}{xxx}

\item{kynthroski}{xxx}

\item{oslaegt}{xxx}

\item{slaegt}{xxx}

\item{lifur}{xxx}

\item{kynfaeri}{xxx}

\item{lengd}{xxx}

\item{leidrett}{xxx}

\item{oracle}{xxx}

\item{kvarnir.col}{xxx}

\item{kvarnir.col.oracle}{xxx}
}
\description{
Read from the table fiskar.kvarnir or the dataframe all.kv.
}

