% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lesa.lengdir.R
\name{lesa.lengdir}
\alias{lesa.lengdir}
\title{Read from the table fiskar.lengdir or the dataframe all.le}
\usage{
lesa.lengdir(synis.id, teg, col.names = c("synis.id", "lengd", "fjoldi"),
  kyn = NULL, kynthroski = NULL, lengd = NULL, leidrett = F,
  oracle = fjolstOracle, lengdir.col = lengdir.col,
  lengdir.col.oracle = lengdir.col.oracle)
}
\arguments{
\item{synis.id}{xxx}

\item{teg}{xxx}

\item{col.names}{xxx}

\item{kyn}{xxx}

\item{kynthroski}{xxx}

\item{lengd}{xxx}

\item{leidrett}{xxx}

\item{oracle}{xxx}

\item{lengdir.col}{xxx}

\item{lengdir.col.oracle}{xxx}
}
\description{
Read from the table fiskar.lengdir or the dataframe all.le
}

