% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lesa.numer.R
\name{lesa.numer}
\alias{lesa.numer}
\title{Read fro the table fiskar.numer or the dataframe all.nu.}
\usage{
lesa.numer(synis.id, teg, vigt = F, col.names = NULL, leidrett = F,
  oracle = fjolstOracle, numer.col = numer.col,
  numer.col.oracle = numer.col.oracle)
}
\arguments{
\item{synis.id}{xxx}

\item{teg}{xxx}

\item{vigt}{xxx}

\item{col.names}{xxx}

\item{leidrett}{xxx}

\item{oracle}{xxx}

\item{numer.col}{xxx}

\item{numer.col.oracle}{xxx}
}
\description{
Read fro the table fiskar.numer or the dataframe all.nu.
}

