% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lesa.stodvar.R
\name{lesa.stodvar}
\alias{lesa.stodvar}
\title{Read from the table fiskar.stodvar or the dataframe stodvar.}
\usage{
lesa.stodvar(col.names = stodvar.col, ar = NULL, veidarfaeri = NULL,
  man = NULL, reitur = NULL, leidangur = NULL, synis.id = NULL,
  oracle = fjolstOracle, stodvar.col = stodvar.col,
  stodvar.col.oracle = stodvar.col.oracle,
  stodvar.col.splus = stodvar.col.splus)
}
\arguments{
\item{col.names}{xxx}

\item{ar}{xxx}

\item{veidarfaeri}{xxx}

\item{man}{xxx}

\item{reitur}{xxx}

\item{leidangur}{xxx}

\item{synis.id}{xxx}

\item{oracle}{xxx}

\item{stodvar.col}{xxx}

\item{stodvar.col.oracle}{xxx}

\item{stodvar.col.splus}{xxx}
}
\description{
Read from the table fiskar.stodvar or the dataframe stodvar.
}

