% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{numer.col.oracle}
\alias{numer.col.oracle}
\title{Columns in the table fiskar.numer in oracle.}
\format{The format is: chr [1:16] "synis_id" "tegund" "fj_maelt"
"fj_kvarnad" ...}
\description{
Columns in the table fiskar.numer in oracle.
}

