% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{numer.col}
\alias{numer.col}
\title{Columns in the dataframe all.nu.}
\format{The format is: chr [1:16] "synis.id" "tegund" "fj.maelt"
"fj.kvarnad" ...}
\description{
Columns in the dataframe all.nu.
}

