\name{oracle.desc}
\alias{oracle.desc}
\title{oracle.desc}
\usage{
  oracle.desc(table, tolower = TRUE, dots = FALSE, ...)
}
\arguments{
  \item{table}{Name of table}

  \item{tolower}{return type}

  \item{dots}{return format}

  \item{...}{Other arguements}
}
\description{
  Some longer text here
}

