\name{oracle.table}
\alias{oracle.table}
\title{oracle.table}
\usage{
  oracle.table(owner = "\%", table = "\%", space = "\%",
    tolower = TRUE, ...)
}
\arguments{
  \item{owner}{Owner}

  \item{table}{Table}

  \item{space}{Space}

  \item{tolower}{Export type}

  \item{...}{Other arguements}
}
\description{
  Some longer text here
}

