% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.sql.R
\name{read.sql}
\alias{read.sql}
\title{Read a file from sql++ with columns seperated by |}
\usage{
read.sql(file, col.names, stop.on.error = T, ...)
}
\arguments{
\item{file}{xxx}

\item{col.names}{xxx}

\item{stop.on.error}{xxx}

\item{...}{xxx}
}
\description{
Read a file from sql++ with columns seperated by |
}

