% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.sqlskipun.R
\name{run.sqlskipun}
\alias{run.sqlskipun}
\title{Internal use.}
\usage{
run.sqlskipun(skipun, listi, colname = "synis_id", txt1 = "and",
  txt2 = "")
}
\arguments{
\item{skipun}{xxx}

\item{listi}{xxx}

\item{colname}{xxx}

\item{txt1}{xxx}

\item{txt2}{xxx}
}
\description{
Internal use.
}

