% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Skala.med.toldum.R
\name{Skala.med.toldum}
\alias{Skala.med.toldum}
\title{Scale length measurements by number counted.}
\usage{
Skala.med.toldum(lengdir, numer, tegund)
}
\arguments{
\item{lengdir}{xxx}

\item{numer}{xxx}

\item{tegund}{xxx}
}
\description{
Scale length measurements by number counted.
}

