% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{skipta.texta}
\alias{skipta.texta}
\title{Translate characters in column names}
\usage{
skipta.texta(txt, char = "_", replchar = ".")
}
\arguments{
\item{txt}{Character vector of inputs}

\item{char}{Character to be replaced}

\item{replchar}{Character to replace with}
}
\value{
A character vector of the same length as \code{txt}.
}
\description{
Translate characters in column names
}
\note{
Could be deprecated, use \code{\link{chartr}} instead.
}
\keyword{manip}

