\name{sql}
\alias{sql}
\title{sql2}
\usage{
  sql(query, tolower = TRUE, dots = TRUE, posix = FALSE,
    encoding = "unknown", useBytes = TRUE,
    stringsAsFactors = FALSE, warn = -1, debug = FALSE,
    ...)
}
\arguments{
  \item{query}{Whatever}

  \item{tolower}{Whatever}

  \item{dots}{Whatever}

  \item{posix}{Whatever}

  \item{encoding}{Whatever}

  \item{useBytes}{Whatever}

  \item{stringsAsFactors}{Whatever}

  \item{warn}{Whatever}

  \item{debug}{Whatever}

  \item{...}{Other things to pass}
}
\description{
  Some description
}
\author{
  AM
}

