% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stodvar.col.oracle}
\alias{stodvar.col.oracle}
\title{Column name in stations file in oracle.}
\format{The format is: chr [1:48] "fiskar.stodvar.synis_id" "leidangur"
"skip" "stod" ...}
\description{
Column name in stations file in oracle.
}

