% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stodvar.col}
\alias{stodvar.col}
\title{Column name in stations frame.}
\format{The format is: chr [1:56] "synis.id" "leidangur" "ar" "man" "dags"}
\description{
Column name in stations frame.
}

