% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stodvar.col.splus}
\alias{stodvar.col.splus}
\title{Column names in fiskar.stodvar with "_" replaced by ".".}
\format{The format is: chr [1:48] "synis.id" "leidangur" "skip" "stod" ...}
\description{
Column names in fiskar.stodvar with "_" replaced by ".".
}

