% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stodvar.std.col.1}
\alias{stodvar.std.col.1}
\title{Subset of culumns in stationsfile.}
\format{The format is: chr [1:25] "synis.id" "leidangur" "skip" "stod" ...}
\description{
Subset of culumns in stationsfile.
}

